+++ src/Equinor.ProCoSys.IPO.MainApi/MainApi/Person/IPersonApiService.cs
@@ -5,6 +5,6 @@ namespace Equinor.ProCoSys.IPO.ForeignApi.MainApi.Person
 {
     public interface IPersonApiService
     {
-        Task<IList<ProCoSysPerson>> GetPersonsByUserGroupAsync(string plant, string searchString, string privilege);
+        Task<IList<ProCoSysPerson>> GetPersonsByUserGroupAsync(string plant, string searchString, string userGroup);
     }
 }
