+++ src/Equinor.ProCoSys.IPO.Query/GetPersonsInUserGroup/GetPersonsInUserGroupQuery.cs
@@ -2,7 +2,7 @@
 using MediatR;
 using ServiceResult;
 
-namespace Equinor.ProCoSys.IPO.Query.GetPersonsByUserName
+namespace Equinor.ProCoSys.IPO.Query.GetPersonsInUserGroup
 {
     public class GetPersonsInUserGroupQuery : IRequest<Result<List<ProCoSysPersonDto>>>
     {
