+++ src/Equinor.ProCoSys.IPO.WebApi/Controllers/Participants/ParticipantsController.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.IPO.Query.GetFunctionalRoles;
-using Equinor.ProCoSys.IPO.Query.GetPersonsByUserName;
+using Equinor.ProCoSys.IPO.Query.GetPersonsInUserGroup;
 using Equinor.ProCoSys.IPO.WebApi.Middleware;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
